requirements.txt
streamlit
requests
duckduckgo-search
rembg
pillow